+++ packages/headless/src/features/insight-search/insight-search-actions.ts
@@ -1,5 +1,10 @@
 import {createAsyncThunk} from '../..';
+import {
+  historyStore,
+  StateNeededByInsightAnalyticsProvider,
+} from '../../api/analytics/insight-analytics';
 import {isErrorResponse} from '../../api/search/search-api-client';
+import {SearchResponseSuccess} from '../../api/search/search/search-response';
 import {
   AsyncThunkInsightOptions,
   InsightAPIClient,
@@ -15,11 +20,7 @@ import {
   SearchSection,
 } from '../../state/state-sections';
 import {requiredNonEmptyString} from '../../utils/validate-payload';
-import {
-  AnalyticsType,
-  makeAnalyticsAction,
-  SearchAction,
-} from '../analytics/analytics-utils';
+import {InsightAction} from '../analytics/analytics-utils';
 import {AnyFacetRequest} from '../facets/generic/interfaces/generic-facet-request';
 import {snapshot} from '../history/history-actions';
 import {extractHistory} from '../history/history-state';
@@ -29,8 +30,13 @@ import {
   FetchQuerySuggestionsThunkReturn,
   StateNeededByQuerySuggest,
 } from '../query-suggest/query-suggest-actions';
+import {getQueryInitialState} from '../query/query-state';
 import {ExecuteSearchThunkReturn} from '../search/search-actions';
-import {logQueryError} from '../search/search-analytics-actions';
+import {getSearchInitialState} from '../search/search-state';
+import {
+  logFetchMoreResults,
+  logQueryError,
+} from './insight-search-analytics-actions';
 
 export type StateNeededByExecuteSearch = ConfigurationSection &
   InsightConfigurationSection &
@@ -56,16 +62,16 @@ const fetchFromAPI = async (
 
 export const executeSearch = createAsyncThunk<
   ExecuteSearchThunkReturn,
-  SearchAction,
+  InsightAction,
   AsyncThunkInsightOptions<StateNeededByExecuteSearch>
 >(
   'search/executeSearch',
   async (
-    analyticsAction: SearchAction,
+    analyticsAction: InsightAction,
     {getState, dispatch, rejectWithValue, extra}
   ) => {
-    /** TODO: We need to dispatch analytics action, but first we have to create InsightClientProvider so the refactor will be available in  https://coveord.atlassian.net/browse/SVCC-2246*/
     const state = getState();
+    addEntryInActionsHistory(state);
     const fetched = await fetchFromAPI(
       extra.apiClient,
       state,
@@ -77,6 +83,18 @@ export const executeSearch = createAsyncThunk<
       return rejectWithValue(fetched.response.error);
     }
 
+    const fetchedResponse = fetched.response.success;
+    analyticsAction(
+      dispatch,
+      () =>
+        getStateAfterResponse(
+          fetched.queryExecuted,
+          fetched.duration,
+          state,
+          fetchedResponse
+        ),
+      extra
+    );
     dispatch(snapshot(extractHistory(getState())));
 
     return {
@@ -122,12 +140,12 @@ export const fetchMoreResults = createAsyncThunk<
 
 export const fetchFacetValues = createAsyncThunk<
   ExecuteSearchThunkReturn,
-  SearchAction,
+  InsightAction,
   AsyncThunkInsightOptions<StateNeededByExecuteSearch>
 >(
   'search/fetchFacetValues',
   async (
-    analyticsAction: SearchAction,
+    analyticsAction: InsightAction,
     {getState, dispatch, rejectWithValue, extra: {apiClient}}
   ) => {
     const state = getState();
@@ -217,12 +235,6 @@ const buildInsightFetchFacetValuesRequest = (
   };
 };
 
-export const logFetchMoreResults = makeAnalyticsAction(
-  'search/logFetchMoreResults',
-  AnalyticsType.Search,
-  (client) => client.logFetchMoreResults()
-);
-
 function getFacetRequests<T extends AnyFacetRequest>(
   requests: Record<string, T> = {}
 ) {
@@ -231,3 +243,41 @@ function getFacetRequests<T extends AnyFacetRequest>(
 
 const getOriginalQuery = (state: StateNeededByExecuteSearch) =>
   state.query?.q !== undefined ? state.query.q : '';
+
+const getStateAfterResponse: (
+  query: string,
+  duration: number,
+  previousState: StateNeededByExecuteSearch,
+  response: SearchResponseSuccess
+) => StateNeededByInsightAnalyticsProvider = (
+  query,
+  duration,
+  previousState,
+  response
+) => ({
+  ...previousState,
+  query: {
+    q: query,
+    enableQuerySyntax:
+      previousState.query?.enableQuerySyntax ??
+      getQueryInitialState().enableQuerySyntax,
+  },
+  search: {
+    ...getSearchInitialState(),
+    duration,
+    response,
+    results: response.results,
+  },
+});
+
+const addEntryInActionsHistory = (state: StateNeededByExecuteSearch) => {
+  if (state.configuration.analytics.enabled) {
+    historyStore.addElement({
+      name: 'Query',
+      ...(state.query?.q && {
+        value: state.query.q,
+      }),
+      time: JSON.stringify(new Date()),
+    });
+  }
+};